// src/generated/models/Stf_strategytacticsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Stf_strategytacticsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_strategytacticsstatecode = keyof typeof Stf_strategytacticsstatecode;
export const Stf_strategytacticsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_strategytacticsstatuscode = keyof typeof Stf_strategytacticsstatuscode;
export const Stf_strategytacticsstf_serviceexecutionmode = {
  1: 'TMS Task',
  2: 'Project'
} as const;
export type Stf_strategytacticsstf_serviceexecutionmode = keyof typeof Stf_strategytacticsstf_serviceexecutionmode;
export const Stf_strategytacticsstf_strategytype = {
  989230000: 'Department Strategy',
  989230001: 'Market Strategy',
  989230002: 'Specialty Strategy',
  989230003: 'Service Strategy'
} as const;
export type Stf_strategytacticsstf_strategytype = keyof typeof Stf_strategytacticsstf_strategytype;
export const Stf_strategytacticsstf_tacticstatus = {
  1: 'Active',
  2: 'Completed',
  3: 'Cancelled'
} as const;
export type Stf_strategytacticsstf_tacticstatus = keyof typeof Stf_strategytacticsstf_tacticstatus;

export interface Stf_strategytacticsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "pm_DriverKPI@odata.bind"?: string;
  "pm_Region@odata.bind"?: string;
  statecode: Stf_strategytacticsstatecode;
  statuscode?: Stf_strategytacticsstatuscode;
  "stf_Assignee@odata.bind"?: string;
  stf_currentbaseline?: number;
  stf_deadline?: string;
  "stf_KPI@odata.bind"?: string;
  stf_neededbudget?: number;
  "stf_Process@odata.bind"?: string;
  stf_serviceexecutionmode?: Stf_strategytacticsstf_serviceexecutionmode;
  "stf_StrategyKPI@odata.bind"?: string;
  stf_strategytacticid: string;
  "stf_TacticCategory@odata.bind"?: string;
  stf_tacticdescription?: string;
  stf_tacticid?: string;
  stf_tacticname?: string;
  stf_tacticstatus?: Stf_strategytacticsstf_tacticstatus;
  stf_target?: number;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Stf_strategytactics extends Stf_strategytacticsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_driverkpiname?: string;
  pm_regionname?: string;
  statecodename?: string;
  statuscodename?: string;
  stf_assigneename?: string;
  stf_assigneeyominame?: string;
  stf_kpiname?: string;
  stf_processname?: string;
  stf_serviceexecutionmodename?: string;
  stf_strategykpiname?: string;
  stf_strategytype?: Stf_strategytacticsstf_strategytype;
  stf_strategytypename?: string;
  stf_tacticcategoryname?: string;
  stf_tacticstatusname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_driverkpi?: object;
  _pm_driverkpi_value?: string;
  pm_region?: object;
  _pm_region_value?: string;
  stf_assignee?: object;
  _stf_assignee_value?: string;
  stf_kpi?: object;
  _stf_kpi_value?: string;
  stf_process?: object;
  _stf_process_value?: string;
  stf_strategykpi?: object;
  _stf_strategykpi_value?: string;
  stf_tacticcategory?: object;
  _stf_tacticcategory_value?: string;
}